# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puraani Jeans APIs",
        "description": "Puraani Jeans API documentation",
        "contact": {
            "email": "developer@puranijeans.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": ""
    },
    "servers": [
        {
            "url": "http://safeblockcom.test/",
            "description": "local"
        }
    ],
    "paths": {
...
        "/v1/closet/{reference}/edit": {
            "post": {
                "tags": [
                    "Closet"
                ],
                "summary": "update closet settings",
                "operationId": "updateCloset",
                "requestBody": {
                    "description": "update closet settings",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "logo": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "banner": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "about": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": []
                    }
                ]
            }
        },
...
        "/v1/product/{productId}": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Product detail",
                "operationId": "getProductDetail",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Get product detail",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "referrer_type": {
                                        "description": "referrer type",
                                        "type": "string",
                                        "example": "Featured Product"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            },
            "client_id": {
                "type": "apiKey",
                "description": "client id",
                "name": "x-client-id",
                "in": "header"
            },
            "client_secret": {
                "type": "apiKey",
                "description": "client secret",
                "name": "x-client-token",
                "in": "header"
            }
        }
    },
...
}